WIP: Attempt to retrieve the temp schema from the connection
Fails when we run, for instance:
PYTHONPATH=`pwd`/src pytest -v tests/databases/test_postgres.py

Because the Airflow Postgres hook is attempting to use all the extra
parameters for setting up the connection

One of the challenges in implementing the feature you describe, is that Postgres expects the connection
extra property only to contain pre-defined Postgres connection fields:
* https://airflow.apache.org/docs/apache-airflow-providers-postgres/stable/connections/postgres.html#howto-connection-postgres
* https://github.com/apache/airflow/blob/59c450ee5425a2d23ef813dbf219cde14df7c85c/airflow/providers/postgres/hooks/postgres.py#L103-L111

So when we give it a connection that has an extra containing another field (e.g. `tmp_schema`), the hook assumes it should be used to establish the connection to Postgres (e.g., as a TTS parameter) and fails to connect.